Added a player manager
The player manager deals with the inventory (towers that are spawned) and interacts vaguely with the game manager. I've also added an optimized healthbar system for the enemies but need to connect it to the attribute system to make it work nicely. Towers have the propensity for an aesthetic mesh, and can be built. Next up I'll have to attach a combiner list to each tower which shows the towers that can be built. It doesn't necessarily have to be a combiner, but maybe it's best if it was. Sorting the list of invertory items to place the currently selected tower on top could afford some measure of control over which tower gets destroyed when building.